Extract FolderNavigatorItem from FolderNavigatorFragment
